Moved history and finding file out of Routing.kt
