+++ script-server/src/main/kotlin/org/geobon/utils/SystemUtils.kt
@@ -18,3 +18,15 @@ fun String.runToText(
         .start().also { it.waitFor(timeoutAmount, timeoutUnit) }
         .inputStream.bufferedReader().readText()
 }.onFailure { it.printStackTrace() }.getOrNull()
+
+
+fun findFilesInFolderByDate(folder:File, fileName: String): List<File> {
+    val process = ProcessBuilder("/bin/bash", "-c", "/bin/bash", "-c", "find $folder -type f -name $fileName -exec stat --format '%Y %n' {} \\; | sort -nr | cut -d' ' -f2-").start()
+    val reader = process.inputStream.bufferedReader()
+
+    val result = mutableListOf<File>()
+    reader.forEachLine { result.add(File(it)) }
+
+    process.waitFor()
+    return result
+}
